backend/app/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.32
sqlmodel==0.0.21
pydantic==2.8.2
pydantic-settings==2.4.0
APScheduler==3.10.4
python-escpos==3.1
python-dotenv==1.0.1